snowflake_poc/requirements.txt
snowflake-connector-python>=3.1.0
sqlalchemy>=2.0.0
pandas>=1.5.0
python-dotenv>=1.0.0
